Added visitor spawning each time the app is opened.

+++ app/src/main/java/uk/co/jakelee/blacksmith/main/VisitorActivity.java
@@ -34,7 +34,7 @@ public class VisitorActivity extends Activity {
         dh = DisplayHelper.getInstance(getApplicationContext());
 
         Intent intent = getIntent();
-        int visitorId = Integer.parseInt(intent.getStringExtra(DisplayHelper.VISITOR_TO_LOAD));
+        Long visitorId = Long.parseLong(intent.getStringExtra(DisplayHelper.VISITOR_TO_LOAD));
 
         if (visitorId > 0) {
             visitor = Visitor.findById(Visitor.class, visitorId);
